+++ src/layers/microscopy-viewer-layer/microscopy-viewer-layer-base.js
@@ -33,15 +33,26 @@ const defaultProps = {
 
 export class MicroscopyViewerLayerBase extends BaseTileLayer {
   constructor(props) {
-    const { sliderValues, colorValues } = props;
+    const { sliderValues, colorValues, channelsOn } = props;
+    const inputColorValues = {};
+    const inputSliderValues = {};
+    Object.keys(colorValues).forEach(channel => {
+      const channelIsOn = channelsOn[channel];
+      inputColorValues[channel] = channelIsOn
+        ? colorValues[channel]
+        : [0, 0, 0];
+      inputSliderValues[channel] = channelIsOn
+        ? sliderValues[channel]
+        : [65535, 65535];
+    });
     const orderedSliderValues = [];
     let orderedColorValues = [];
-    Object.keys(sliderValues)
+    Object.keys(inputSliderValues)
       .sort()
-      .forEach(key => orderedSliderValues.push(sliderValues[key]));
-    Object.keys(colorValues)
+      .forEach(key => orderedSliderValues.push(inputSliderValues[key]));
+    Object.keys(inputColorValues)
       .sort()
-      .forEach(key => orderedColorValues.push(colorValues[key]));
+      .forEach(key => orderedColorValues.push(inputColorValues[key]));
     const diffSliders = 6 - orderedSliderValues.length;
     for (let i = 0; i < diffSliders; i += 1) {
       orderedSliderValues.push([0, 65535]);